Move hermes to official dir
