+++ official/hermes/cmd/hermes_start.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/ignite/apps/hermes/pkg/hermes"
+	"github.com/ignite/apps/official/hermes/pkg/hermes"
 )
 
 // NewHermesStart start the hermes relayer.
